Basic Vector functionality
Managed to make all of the allocations work correctly without links,
using the allocator supplied as an optional template argument.

A custom iterator is also implemented, and I'll create a new one soon
that inherits from the existing iterator, with any modifying
functionality overridden so that the iterator is const.

Currently experiencing some difficulties however in getting the rotate()
function from the standard algorithm library working, which is annoying,
since reverse() and sort() both appear to work just fine.