Removed commented nested iterator
Now that the RandomAccessIterator Class is functioning correctly, I've
removed the old implementation of a nested Class iterator inside of the
Vector Class.

I also removed the const_pointer and const_reference typedefs inside of
the RandomAccessIterator Class, as using const T* and const T& is sufficient.

My next goal is to create a ReverseIterator Class, which should be less
work than the RandomAccessIterator, all I'll do is nest an iterator
inside of the ReverseIterator (which will be the RandomAccessIterator or
a new BidirectionalIterator Class, typenamed to Iter), and then write
new operator overloads for all of the relevant maths involved;