Reverse iterator functional
I'm pretty sure that the basic functionality of the reverse iterator is
in place, after getting a bunce of off-by-one errors for a while.

I'll have to perform some additional checks of course, I'll make sure
that all of the assignments that should be possible are possible
(assigning a reverse iterator from a normal iterator for example, is
that possible?).

Then finally, after all of the difficult work, I can move on to making
the vector member functions, which should be a relaxing break after all
of this tough iterator work.