I think Reverse Iterators are complete
Comparisons are finally added for the ReverseIterator class, using the
same trick as with the RandAccessIterator.

Tests are extensive, but all tests only print to the screen and are not
automated. This may be good enough however, I think I'll leave it like
this for now.

One thing that I might change however is that there is a test file that
is something like 100 lines long, or bigger. I think it would be a good
idea to divide this up a little bit.