Very odd behaviour in insert() function for vector
Going to have to try to figure this one out tomorrow.
I think the algorithm for both insert() overloads is correct, but
compilation is strange as the value_type and the size_type types can get
messed up during compilation - not just for my ft::vector, but for
std::vector too.

It may be that you have to do some strange type of casting to get it to
work properly, and if I can find what that casting is, maybe it will
work just fine for mine as well.

That said, I'm pretty certain that SFINAE is going to be required for
the final overload of insert(), which I'm sure I can really look forward
to.