Fixed inefficiency/discrepancy with std::vector
I was doing the copy construction (well, the overload of the = operator)
in a lazy way previously - simply re-reserving the new memory capacity,
and then filling the new vector with the values from the vector to be
copied.

Now, I think the algorithms for the copy should be more efficient, only
constructing/destructing/allocating/deallocating as much as it should.

Also, the copy construction does not assign equal capacity -> if the
vector being copied into has a larger capacity already, then it will
retain this increased capacity in std::vector. ft::vector now does this
also.