Bug squashing
Found a very small bug inside of the insert() function, the one where
only a single element is inserted. I was failing to call destroy on the value at
the space where I was about to insert the new one - fixed.

New error located in the erase() function, I'll have to try to find time
to fix it.