Commit before removing construct for entire cap
At the moment, I'm constructing values for the entire capacity of a
vector - using default values for the sections of the vector that are
not within the vectors size. I think this could be the incorrect way of
doing it, so I'm going to go through the code and remove unnecessary
constructs, and remove unnecessary destroys.

Making this commit so that I can revert if I'm wrong about this.