List is well begun
Basic list logic is setup.

List relies on a nested 'node' class, private from calling code.
The nested node class contains the values that are placed into the list,
and with each value, also contains two pointers, one for the previous
value of the list and one for the next value of the list.

The most basic list functions have been created (clear, push_back,
push_front, etc.). Valgrind has been used to check that the list manages
memory effectively.

A new iterator class ListIterator has been created for the list. As it
is only a bidirectional iterator, it doesn't have the same level of
functionality as the random access iterator for the vector, and there is
no need for > < >= <= comparisons. I think that all of the required
functionality for the bidirectional iterator is in place, but I can't be
sure yet.

I also am yet to try using the bidirectional iterator with the
ReverseIterator. Hopefully it's as simple as placing the ListIterator
in as a template parameter for the ReverseIterator in the list, but as
there is more functionality present in the ReverseIterator it may not
work out so easily.