Change needed for list implementation
I've found something really cool concerning the application of an
iterator in a std::list. I was confused as to how you could continue to
add to an iterator that began from end() or rend(), since the linked
list would eventually become null.

After some testing however, I found that the linked list appears to form
a sort of circle, with the "tail" connecting to the "head" via a DUMMY
NODE. Interestingly, this Dummy node appears to store the size of the
list.

EXTRA NOTE:
	After one more test, I'm seeing that there is the same looping
	effect in a std::list with a std::string, BUT that it does not
	store the size in the dummy node, the dummy node simply causes a
	segmentation fault upon dereferencing.

I think I will go for this dummy node implementation, with a _val that is
never constructed.

ADDITIONAL IMPORTANT NOTE:
	I missed a couple of the constructors in the vector class, I
	need to go back and complete them.