sort() done
Followed a guide to the algorithm much much better this time than
previously - using node pointers instead of iterators, which was much
more tractable.

It was really important, simple, and pretty obvious on reflection, to
use a comparison to the dummy node as a base case for the recursive
quicksort function.

It's not a introspective sort, but it should be good enough for the
purpose of this project.