Wrote ft::stack
Was just as trivial as expected.

I should write some tests of course for the stack, which will probably
take longer than creating the stack class in the first place. As there
aren't a large number of member functions to test however, it shouldn't
be too difficult a procedure.

Now only 2 containers remaining (but one of them is map...)!