Minor edit to vector resize
I happened to notice that the capacity for std::vector varies in very
odd ways, it seems that if the new size from a resize is greater than
capacity, but less than size * 2, then the new capacity becomes size *
2.

I have changed my vector to match this behaviour, but it's also a very
minor thing. If I've missed some odd behaviour of the std::vector, I
think that that's justifiable, it doesn't matter if the capacity doesn't
match up exactly, as long as the vector is allocating memory dynamically
as it should.