Map is begun
I haven't worked on this project for around 3 weeks, I'll try to
reacquaint myself with it today.

As far as I can remember, the vector and list are completed - I could
potentially rewrite the tests with a real unit test, and I might use
GoogleTest (which I've started using at work) to do this.

The primary objective however, is to write the ft::map.
I don't think this will be very difficult, but it will probably take a
bunch of time.

I also saw that apparently ft::map does require comparison operators,
even though it's not on cplusplus.com.
Annoying, but good to know.