Beginning to write Google Tests
Should make tests much nicer, and its just a good thing to expand on my
Unit testing experience further, so I'm going to try to create an
expansive Unit Test suite for the ft_containers rather than just the
printed out tests I have at the moment.

It was frustrating trying to mix generic programming with googletest,
and I was getting strange errors for a long time, but after a bunch of
fiddling I've apparently stumbled onto the right way of doing it.