Bug fix and Example class added
I added an Example class to the tests/ directory. Nothing fancy, just to
show that the ft containers work just fine with custom objects as well as ints
and std::strings.

I changed the test for the the max_size, since the values come out very
slighly different, I now round them to the nearest 100.

I found an error in the erase() function for the ft::vector, where I was
constructing values I shouldn't be, from uninitialized memory. Fixed
now, just had to add an if statement before a _a.construct() line.