List Google Tests completed - base() added
I had forgotten to add a base() function to the reverse iterator for
the list, which I've now noticed. It's failing the test at the moment,
but I'm sure I can fix it shortly.

All but 5 of the list tests are failing currently, although I will need
to also create the tests for the function unique to the list, that don't
exist in ft::vector.