Identified improvement in list with splice tests
There's an odd thing about the splice function, if you try to splice
using an incorrect iterator, to the list you're calling the method
function with for example, then the size of the list will increase,
while the list doesn't actually get any bigger.

While this is erroneous usage of splice, it was also causing double
frees in my list implementation due to my clear() method using _size to
determine how much to destroy/free. Now the size is reduced by clear(),
but the actually determinant for how much to free is the actual nodes
inside of the list. Double free is impossible as a result.