List tests completed
Added the tests that are unique to list, and not in vector.

No further changes were required to ft::list to pass all tests. I
thought I may need SFINAE for the unique function briefly, but it turns
out I just needed to determine the correct Predicate function inside of
the test function itself, which is done by initializing a function
pointer to the correct function before passing the function pointer as a
parameter.

On to map, or maybe tests for the stack.