Non-functional commit
Changing the return type of insert.
Also adding the logic so that a pair with an identical key isn't added
to the red black binary tree.

At the moment I think this is possible, and it appears to have broken
the iteration as well, although the iteration is already not great.

I'm rethinking the iteration as well, I don't actually believe that it
would be necessary to maintain a _previous pointer in the iterator, I
just need to be smarter in my solution, it's really not that hard.