Finally got second ft::map::insert() variant
The ft::map::insert() variant that uses a hint to increase efficiency of
insertion was a real pain, as the std::map will still retain a correct
order when the hint is incorrect, it simply reverts to using the
standard method of insertion.

My code to implement this is super ugly, but I've confirmed that the
efficient insertion will be used if the hint is correct, and otherwise
the standard insertion will take place. If I have time I'll clean the
code up but I don't really think it's worth it.