Added extra map constructors
The range and copy constructors for ft::map have now been added.

It was trivial, except for a slight oversight where I overlooked the
initialization list when using the copy constructor. This was causing
the ft::map::clear() function to break entirely.

Fixed, tests written, and all passing (with valgrind).