Absolute hell of a commit - const conversions
Conversion from a non-const iterator for ft::map to a const iterator is
now possible, although not yet tested.

This was absolute hell for the longest time, I was pretty stupid to not
test what 'const_type' inside the iterator actually meant, but by
finally realising i could test it's type by attempting to 'cout' it, I
saw there was actually a mapped_type, instead of a pointer to
ft::pair<key, mapped_type>, which is what I needed.

After a slightly hacky solution (involving adding a new member type to
the ft::map::node), I've finally reached compilation.

This was a really stupid thing to mess up, but I should bear in mind in
future that especially when working with generics, types may not be the
type you think they are - and testing with cout is a super easy way to
verify type.