While creating ft::map tests, found bug in list
I had accidentally missed testing reverse bidirectional iterator
relational operators in both directions, meaning a const reverse bi
iterator could be compared to a non-const reverse bi iterator, by way of
implicit conversion of a non-const to const.

It could not however take place in the reverse direction.

To fix, I move the comparison operator overload outside of the
ReverseBiIterator class, made it a generic to take two different types
of iterator (const and non-const), made it friend so it could access the
underlying pointer, and compare this pointer.

I'll add a few tests to the list iterator tests, and then complete the
tests for ft::map::iterator comparisons.