I should have added a dummy node to ft::map
I'm feeling a little stupid for not foreseeing this as a serious issue,
but I'm having problems with the iterators due to there not being a
dummy node in ft::map.

I'll try to amend this now, but it's probably going to be a little
messy, as it won't play well with the existing logic I have for the map,
especially with the reshuffling.