Broke everything
I'm trying to hack in a dummy node into the ft::map, it now appears to
be working well for basic iteration, but there's some issue occurring in
the insert_hint tests.

(I also fixed a bug in the insert hint tests, if a node was unable to be
added (due to existing in the map already) then it wasn't being
destroyed - fixed now).

The dummy implementation is a little weird, it points to the left most
and rightmost nodes of the map, and they point to the dummy. But
whenever an insert takes place, the dummy disconnects itself before
insert logic, then reconnects itself again. All of this is much easier
said than done.

When I implement deletion, then I'm going to have to include additional
checks to keep the dummy in order.