Extensive modification of tree logic complete
I've completed the addition of the dummy node to the map - completed
with additions to iteration logic to work correctly.

My iteration does have undefined behaviour that is slightly different
from the std::map - namely, if you increment the end iterator of an
ft::map, it moves to the dummy, and then back to the final node again,
and loops over just these two nodes.

My ft::map::iterator is more similar to the linked list iterator, if you
add to the map::end() iterator, it will loop over to the begin(), and
iterate over the map a second time.

The insert with hint function is working correctly as well, the method
I've chosen to implement the dummy is to disconnect before insertion,
and then reconnect afterwards.

I'm a little worried that there may be aspects I've missed, I hope that
my testing is thorough enough to cover all possibilities.

When I add implementation for deletion, I'm going to have to add
additional logic to make sure that the dummy node remains valid.