Implementation of delete node begun
I've added the logic that I think is necessary for a BST-style deletion,
but I haven't yet added the logic for recoloring the tree, which could
definitely lead to breaks in logic if not fixed.

After that's added, I'm sure some tests will expose how incorrect and
broken everything is.