Bug fixes
I thought ft::map::erase() was working before, but I hadn't tested it
thoroughly enough - only tests using the position of the very first
element were included.

After adding a test that would remove the 3 element of the map, I
encountered bugs due to my implementation of the swap function between
two node pointers in the map.

I've now changed it so that instead of swapping the nodes entirely, the
values are simply swapped. This required a const_cast, but makes the
code much more tractable.

Tests are passing now, but I should probably add another test using an
iterator to the right hand side of the map.

A lot of debug code is still present, shall be removed later.