erase() functions for map complete
I added the find() and erase functions to map().
The erase by range function in map is particularly ugly, due to the way
the map was designed initially.

It's gross, but it's working well, and passing all tests with no
valgrind errors.

Given how close I am to the end of the project, I don't believe its
worth rewriting.

Only a few more functions left.