Added tests for stack. Fixed bug in vector
While adding tests for the stack, I discovered a bug in ft::vector that
must have existed for some time, that somehow slipped through the tests.
The = operator overload for ft::vector would allocate a new region of
memory for the new vector based on the SIZE of the other vector, not the
capacity.

This was relevant for implicit copy construction of vectors, and found
while creating basic stack tests.

Bug fixed, and tests added for all stack functions (there weren't many).

Next target is to create ft::queue (should be simple), write some tests
(basically the same as the stack tests), and then submit the project.